requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.14
djangorestframework==3.13.1
idna==3.3
lxml==4.9.1
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.9
